+++ NEWS.md
@@ -7,6 +7,8 @@
 
 * In `grepl()`, the argument `fixed` is now used correctly (thanks @gernophil
   for the report, #223).
+  
+* `if_else()` and `ifelse()` now work when using named arguments (#224).
 
 # tidypolars 0.14.0